Fixed some gcc warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/firebird/code/firebird/trunk@58810 65644016-39b1-43b1-bf79-96bc8fe82c15

+++ src/lock/print.cpp
@@ -657,7 +657,7 @@ int CLIB_ROUTINE main( int argc, char *argv[])
 		}
 
 		LOCK_header = (lhb*)(UCHAR*) buffer;
-		const size_t bytes_read = read(fd, LOCK_header, file_stat.st_size);
+		const int bytes_read = read(fd, LOCK_header, file_stat.st_size);
 		close(fd);
 
 		if (bytes_read != file_stat.st_size)
